+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## 0.7.0-dev
-- Removed id as own element variable it's part of attributes now
+- Removed id as own element variable, it's part of attributes now
 - Removed String allocations for better performance
+- Added formatted html output
 
 ## 0.6.0
 - Added github ci/cd workflows